+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.1
+// swift-tools-version:5.2
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -20,13 +20,11 @@ let package = Package(
         .package(url: "https://github.com/OpenKitten/MongoKitten.git", from: "6.0.0"),
     ],
     targets: [
-        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
-        // Targets can depend on other targets in this package, and on products in packages which this package depends on.
         .target(
             name: "FluentMongoDriver",
             dependencies: ["FluentKit", "MongoKitten"]),
-        .testTarget(
-            name: "FluentMongoDriverTests",
-            dependencies: ["FluentMongoDriver", "FluentBenchmark"]),
+//        .testTarget(
+//            name: "FluentMongoDriverTests",
+//            dependencies: ["FluentMongoDriver", "FluentBenchmark"]),
     ]
 )
