--- Tests/LinuxMain.swift
@@ -1,6 +1,0 @@
-import XCTest
-import FluentMongoDriver
-
-var tests = [XCTestCaseEntry]()
-tests += FluentMongoDriverTests.allTests()
-XCTMain(tests)
